backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "fin-agent-backend"
version = "0.1.0"
description = "Financial Asset Q&A Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.40",
    "requests>=2.31.0",
    "chromadb>=0.5.0",
    "openai>=1.30.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "finnhub-python>=2.4.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
